refactor : separate class
separate some method using in authentication from userService
