+++ src/main/java/com/fastcampus/scheduling/_core/security/filter/CustomAuthenticationProvider.java
@@ -3,15 +3,14 @@ package com.fastcampus.scheduling._core.security.filter;
 import static com.fastcampus.scheduling._core.errors.ErrorMessage.MISMATCH_SIGN_IN_INFO;
 
 import com.fastcampus.scheduling._core.errors.exception.Exception400;
-import com.fastcampus.scheduling.user.service.UserService;
-import javax.annotation.Resource;
+import com.fastcampus.scheduling.user.service.CustomDetailService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
 @Slf4j
@@ -19,17 +18,15 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class CustomAuthenticationProvider implements org.springframework.security.authentication.AuthenticationProvider {
 
-    @Resource
-    private final UserService userService;
-    private final BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+    private final CustomDetailService customDetailService;
+    private final PasswordEncoder passwordEncoder;
 
     @Override
     public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-        UsernamePasswordAuthenticationToken authenticationToken = (UsernamePasswordAuthenticationToken) authentication;
-        String userEmail = (String) authenticationToken.getPrincipal();
-        String userPassword = (String) authenticationToken.getCredentials();
+        String userEmail = authentication.getPrincipal().toString();
+        String userPassword = authentication.getCredentials().toString();
 
-        UserDetails userDetails = userService.loadUserByUsername(userEmail);
+        UserDetails userDetails = customDetailService.loadUserByUsername(userEmail);
 
         if (!passwordEncoder.matches(userPassword, userDetails.getPassword())) {
             log.warn("password Not Match From matches");
