+++ src/main/java/com/fastcampus/scheduling/user/service/UserService.java
@@ -4,22 +4,15 @@ import com.fastcampus.scheduling._core.common.Constants;
 import com.fastcampus.scheduling._core.errors.ErrorMessage;
 import com.fastcampus.scheduling._core.errors.exception.Exception400;
 import com.fastcampus.scheduling._core.errors.exception.Exception401;
-import com.fastcampus.scheduling._core.security.dto.UserPrincipal;
 import com.fastcampus.scheduling.admin.dto.AdminRequest;
 import com.fastcampus.scheduling.user.common.Position;
 import com.fastcampus.scheduling.user.dto.UserRequest;
 import com.fastcampus.scheduling.user.dto.UserRequest.UpdateDTO;
 import com.fastcampus.scheduling.user.model.User;
 import com.fastcampus.scheduling.user.repository.UserRepository;
-import java.util.ArrayList;
-import java.util.Collection;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.OptimisticLockingFailureException;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -28,28 +21,10 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class UserService implements UserDetailsService {
+public class UserService {
 
 		private final UserRepository userRepository;
 		private final PasswordEncoder passwordEncoder;
-
-		@Override
-		public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-			User user =  userRepository.findByUserEmail(username)
-					.orElseThrow(() -> new UsernameNotFoundException(ErrorMessage.MISMATCH_SIGN_IN_INFO));
-
-			Collection<GrantedAuthority> authorities = getAuthorities(user);
-
-			return new UserPrincipal(user, authorities);
-		}
-
-		public Collection<GrantedAuthority> getAuthorities(User user) {
-			Collection<GrantedAuthority> authorities = new ArrayList<>();
-			String authority = "ROLE_";
-			authority += user.getPosition() == Position.MANAGER ? "MANAGER" : "USER";
-			authorities.add(new SimpleGrantedAuthority(authority));
-			return authorities;
-		}
 
 		@Transactional(readOnly = true)
 		public User findByUserId(Long userId) {
